+++ src/edge.hpp
@@ -14,7 +14,7 @@
 
 class Edge;
 std::unique_ptr<Edge> createEdge(uint32_t begin_node_id, uint32_t end_node_id,
-    uint32_t label, int32_t weight = 1);
+    uint32_t label, float weight = 1.0);
 
 class Edge {
 public:
@@ -33,28 +33,28 @@ public:
         return sequence_labels_;
     }
 
-    const std::vector<int32_t>& sequence_weights() const {
+    const std::vector<float>& sequence_weights() const {
         return sequence_weights_;
     }
 
-    int32_t total_weight() const {
+    float total_weight() const {
         return total_weight_;
     }
 
-    void add_sequence(uint32_t label, int32_t weight = 1);
+    void add_sequence(uint32_t label, float weight = 1.0);
 
     friend std::unique_ptr<Edge> createEdge(uint32_t begin_node_id,
-        uint32_t end_node_id, uint32_t label, int32_t weight);
+        uint32_t end_node_id, uint32_t label, float weight);
 
 private:
 
-    Edge(uint32_t begin_node_id, uint32_t end_node_id, uint32_t label, int32_t weight);
+    Edge(uint32_t begin_node_id, uint32_t end_node_id, uint32_t label, float weight);
     Edge(const Edge&) = delete;
     const Edge& operator=(const Edge&) = delete;
 
     uint32_t begin_node_id_;
     uint32_t end_node_id_;
     std::vector<uint32_t> sequence_labels_;
-    std::vector<int32_t> sequence_weights_;
-    int32_t total_weight_;
+    std::vector<float> sequence_weights_;
+    float total_weight_;
 };
